WIP Add support for bundles
This commit adds support for checking the state of a bundle.
A bundle is a container managed by pacemaker and can be a plain
container being started/stopped by pacemaker *or* it can be a
container that has pacemaker_remote and an OCF resource agent running
inside.

We use the same strategy as for other resources: we calculated
the expected count (which is the masters count for M/S resources
and the replicas count in all other cases) and match it against
the running resources count.

Change-Id: I1a9c82ab3056c4d45d63ce3c43b2e59cc6cfd6cb
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>