Change small nits in the pacemaker_is_active module.
The type for the input parameters wasn't defined
and as a consequence, when runnign the modules a
WARNING message was being printed out:
The value 5 (type int) in a string field was converted
to u'5' (type string). If this does not look like what
you expect, quote the entire value to ensure it does
not change.

Change-Id: I8c9a23faced4bc6cf399ce3e896f6127ecee09c2