Pass --wait option only on enable,disable,restart.
The pacemaker resource Ansible module has a wide set
of states which are being passed into the execution of the
pcs resource command. However, not all the options supported
work fine with the --wait argument.

This patch appends the --wait option in the case the state
passed into the module is enable, disable or restart.

Resolves: rhbz#1805841

Change-Id: Ie1b1027c5f861cbfa79583524017642c5bca1b17