Make get_cluster_status() more robust
get_cluster_status() is currently grepping for a specific string
in order to figure out if the the cluster is online or not.
This is not very robust for at least two reasons:
1) The returned error string might change over time or due to locales
2) pcs cluster status returns ok even when the node has not yet fully
   joined the cluster.

Especially the latter is problematic because we can end up tweaking
the CIB before the node has fully joined the cluster and this is not
always catered for in pcmk (see related-bug).

We add a simple additional check to make sure the cluster is really
up. Namely we call 'crm_node --quorum' which will output '1' if the
node is part of the quorate partition.

And we also stop grepping for a specific message and instead rely on
the error code returned by pcs.

This should give us sufficient guarantees that the cluster is online
and ready to accept any commands on the CIB.

Tested as follows in a Train / IHA / Composable HA control-plane:
- Run a full deployment with this patch
- Did an overcloud reboot
- Launched a minor update

Related-Bug: rhbz#1961857

Change-Id: Ic550450901010ff4560a62597d9aaebd59ff5357