Fix: use the correct command to get a resource
The pacemaker_resource module uses the `pcs resource status <resource>`
tho fetch the status of a resource, but that command, originally removed
in pcs 0.10.1, was "temporarily" readded and deprecated in pcs 0.10.2 [1]
and finally removed in pcs 0.11.1.

This patch uses the current command with the same effect (it shows
the resource if available and 0 as return code, but returns a non-0
return code if the resource does not exist), and it should work
even on older pcs 0.10.x versions.

Thanks Alan Bishop for the investigation.

[1] https://github.com/ClusterLabs/pcs/commit/93da15a7615e8a83cbc220c8ad03197dc702df04
[2] https://github.com/ClusterLabs/pcs/commit/240faf0c5dda0321ccae6e34d51139626f9b30a6

Resolves: rhbz#2252464
Change-Id: Iba0b69e8bced7eb027ac9286a729bd5fadf15294